requirements.txt
alembic==1.4.2
appdirs==1.4.3
asn1crypto==1.3.0
attrs==18.2.0
Babel==2.8.0
bcrypt==3.1.7
Beaker==1.11.0
beautifulsoup4==4.9.0
blinker==1.4
borgbackup==1.1.11
boto==2.48.0
breezy==3.0.2
Brotli==1.0.7
btrfsutil==1.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
configobj==5.0.6
contextlib2==0.6.0.post1
cryptography==2.9.2
decorator==4.4.2
diskcache==2.9.0
distlib==0.3.0
distro==1.5.0
dulwich==0.19.16
extras==1.0.0
fastimport==0.9.8
filelock==3.0.12
fixtures==3.0.0
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==3.3.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
greenlet==0.4.15
h11==0.9.0
h2==3.2.0
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.2.0
idna==2.9
imdbpie==5.6.4
importlib-metadata==1.5.1
isc==2.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kaitaistruct==0.8
ldap3==2.7
lensfun==0.3.95
llfuse==1.3.6
lockfile==0.12.2
lxml==4.5.0
Mako==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
meson==0.54.1
mitmproxy==5.1.1
msgpack==1.0.0
numpy==1.18.3
ordered-set==3.1.1
packaging==20.3
pandocfilters==1.4.2
panflute==1.12.4
paramiko==2.7.1
passlib==1.7.2
patiencediff==0.1.0
pbr==5.4.4
pep517==0.8.2
Pillow==6.2.1
ply==3.11
progress==1.5
protobuf==3.11.4
psutil==5.7.0
psycopg2==2.8.5
publicsuffix2==2.20191221
pulsemixer==1.5.0
pwquality==1.4.2
pyasn1==0.4.8
PyAudio==0.2.11
pycairo==1.19.1
pycountry==19.8.18
pycparser==2.20
pycryptodome==3.9.7
PyGObject==3.36.0
PyNaCl==1.3.0
pynvim==0.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.7.0
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQtWebEngine==5.14.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-libtorrent==1.2.5
python-mimeparse==1.6.0
pytoml==0.1.21
pytz==2019.3
PyYAML==5.3.1
ranger-fm==1.9.3
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Send2Trash==1.5.0
shutilwhich==1.1.0
simplejson==3.17.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
speaklater==1.3
SQLAlchemy==1.3.16
sqlparse==0.3.1
sshtunnel==0.1.5
streamlink==1.3.1
team==1.0
termcolor==1.1.0
testtools==2.4.0
tldr==0.5
toml==0.10.0
tornado==6.0.4
trans==2.1.0
urllib3==1.25.9
urwid==2.1.0
virtualenv==20.0.18
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wsproto==0.15.0
WTForms==2.2.1
youtube-dl==2020.3.24
zipp==3.1.0
zstandard==0.13.0
